+++ include/xmol/v1/proxy/smart/AtomSmartSelection.h
@@ -91,7 +91,7 @@ public:
     return m_selection.empty();
   }
 
-  AtomRef& operator[](int i) {
+  AtomRef& operator[](size_t i) {
     check_precondition("operator[]()");
     return m_selection[i];
   }
