+++ include/xmol/v1/proxy/smart/ResidueSmartSpan.h
@@ -51,7 +51,7 @@ public:
     return m_span.empty();
   }
 
-  ResidueRef operator[](int i) {
+  ResidueRef operator[](size_t i) {
     check_precondition("operator[]()");
     return m_span[i];
   }
